feat(notifications): add stimulus controller for notifications and refactor form fill
Refactor form fill controller to use new notification controller instead of inline flash messages
