+++ app/views/form_fills/show.html.erb
@@ -128,7 +128,7 @@
     }
   </style>
 <% end %>
-<div class="min-h-screen bg-gradient-to-br from-slate-900 via-blue-900 to-indigo-900">
+<div class="min-h-screen bg-gradient-to-br from-slate-900 via-blue-900 to-indigo-900" data-controller="notification" data-action="showNotification@window->notification#showNotification">
   <div class="container mx-auto px-4 py-6 lg:py-12">
     <div class="flex flex-col sm:flex-row sm:items-center sm:justify-between mb-8 animate-fade-in">
       <div class="mb-4 sm:mb-0">
